Create a URL for the given tablename with optional filtering

This function creates a URL for the specified tablename by combining the base URL and the endpoint for the tablename. If provided, it also appends filter strings to the URL. If a filter list is provided, a user authentication token is required.

Usage

create_url(
  tablename,
  base_url = database_info$base_url,
  filter_list = NULL,
  token = NULL
)

Arguments

tablename

The name of the table for which to create the URL.

base_url

The base URL of the API (default: database_info$base_url).

filter_list

A list of filter strings to append to the URL (default: NULL).

token

The user authentication token required for adding filter strings (default: NULL).

Value

The complete URL for the specified tablename, with filter strings appended if provided.

See also

Examples

if (FALSE) {
# Without filters
create_url("ccexperiment")

# With filters
create_url("ccexperiment",
            filter_list = list("id=1", "category=sample"),
            token = "your_auth_token")
}